requirements.txt
pysus
pandas
psycopg2-binary
sqlalchemy
requests
python-dotenv
openai
streamlit
plotly